pyproject.toml
[project]
name = "oubon_mailbot"
version = "0.1.0"
description = "Oubon MailBot: FastAPI Gmail automation with AI replies, routing, and alerts."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.34.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.141.0",
    "SQLAlchemy>=2.0.32",
    "alembic>=1.13.2",
    "openai>=1.42.0",
    "requests>=2.32.3",
    "jinja2>=3.1.4",
    "slack_sdk>=3.33.1",
    "python-amazon-paapi>=5.0.0",
    "tenacity>=9.0.0",
    "pytrends>=4.9.2",
    "pandas>=2.2.2",
]

[project.optional-dependencies]
worker = ["redis>=5.0.8", "rq>=1.16.2"]
gcp = ["google-cloud-pubsub>=2.21.6"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
    "httpx>=0.27.0",
    "ruff>=0.6.9",
    "responses==0.25.0",
    "pytest-mock>=3.14.0",
]
